% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{cart_add_filter}
\alias{cart_add_filter}
\title{Add a filter to the cart}
\usage{
cart_add_filter(cart, filter, id = NULL)
}
\arguments{
\item{cart}{An omop_cart}

\item{filter}{An omop_filter or omop_filter_group object}

\item{id}{Character or NULL; filter ID (auto-generated if NULL)}
}
\value{
The modified cart
}
\description{
Add a filter to the cart
}
