% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{cart_add_output}
\alias{cart_add_output}
\title{Add an output specification to the cart}
\usage{
cart_add_output(cart, output)
}
\arguments{
\item{cart}{An omop_cart}

\item{output}{An omop_output object}
}
\value{
The modified cart
}
\description{
Add an output specification to the cart
}
