% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{cart_add_population}
\alias{cart_add_population}
\title{Add a population to the cart}
\usage{
cart_add_population(cart, population)
}
\arguments{
\item{cart}{An omop_cart}

\item{population}{An omop_population object}
}
\value{
The modified cart
}
\description{
Add a population to the cart
}
