% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{cart_add_variable}
\alias{cart_add_variable}
\title{Add a variable to the cart}
\usage{
cart_add_variable(cart, variable = NULL, ...)
}
\arguments{
\item{cart}{An omop_cart}

\item{variable}{An omop_variable object, or NULL to construct from ...}

\item{...}{If variable is NULL, arguments passed to omop_variable()}
}
\value{
The modified cart
}
\description{
Add a variable to the cart
}
