% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{cart_import_json}
\alias{cart_import_json}
\title{Import a cart from JSON}
\usage{
cart_import_json(json)
}
\arguments{
\item{json}{Character; JSON string or file path}
}
\value{
An omop_cart object
}
\description{
Import a cart from JSON
}
