% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{cart_remove_filter}
\alias{cart_remove_filter}
\title{Remove a filter from the cart}
\usage{
cart_remove_filter(cart, id)
}
\arguments{
\item{cart}{An omop_cart}

\item{id}{Character; filter ID to remove}
}
\value{
The modified cart
}
\description{
Remove a filter from the cart
}
