% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{cart_remove_output}
\alias{cart_remove_output}
\title{Remove an output specification from the cart}
\usage{
cart_remove_output(cart, name)
}
\arguments{
\item{cart}{An omop_cart}

\item{name}{Character; output name to remove}
}
\value{
The modified cart
}
\description{
Remove an output specification from the cart
}
