% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{cart_remove_population}
\alias{cart_remove_population}
\title{Remove a population from the cart}
\usage{
cart_remove_population(cart, id)
}
\arguments{
\item{cart}{An omop_cart}

\item{id}{Character; population ID to remove}
}
\value{
The modified cart
}
\description{
Remove a population from the cart
}
