% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{cart_remove_variable}
\alias{cart_remove_variable}
\title{Remove a variable from the cart}
\usage{
cart_remove_variable(cart, name)
}
\arguments{
\item{cart}{An omop_cart}

\item{name}{Character; variable name to remove}
}
\value{
The modified cart
}
\description{
Remove a variable from the cart
}
