% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{cart_to_code}
\alias{cart_to_code}
\title{Generate reproducible R code from a cart}
\usage{
cart_to_code(cart)
}
\arguments{
\item{cart}{An omop_cart object}
}
\value{
Character string of R code
}
\description{
Produces minimal R code that recreates the cart. Does not include
library() calls or header comments.
}
