% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.build_code}
\alias{.build_code}
\title{Build an R code string for a function call}
\usage{
.build_code(fn_name, ...)
}
\arguments{
\item{fn_name}{Character; fully qualified function name}

\item{...}{Named arguments to include in the call}
}
\value{
Character string of the R call
}
\description{
Build an R code string for a function call
}
\keyword{internal}
