% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{.circe_codeset_details}
\alias{.circe_codeset_details}
\title{Resolve and validate a Circe CodesetId}
\usage{
.circe_codeset_details(codeset_id, concept_sets)
}
\description{
Resolve and validate a Circe CodesetId
}
\keyword{internal}
