% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{.circe_criteria_to_filter}
\alias{.circe_criteria_to_filter}
\title{Convert one Circe occurrence criteria back into a recipe filter}
\usage{
.circe_criteria_to_filter(crit, concept_sets)
}
\description{
Convert one Circe occurrence criteria back into a recipe filter
}
\keyword{internal}
