% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{.circe_demographic_from_filter}
\alias{.circe_demographic_from_filter}
\title{Append a demographic criteria (sex / age) for a recipe filter}
\usage{
.circe_demographic_from_filter(f)
}
\description{
Append a demographic criteria (sex / age) for a recipe filter
}
\keyword{internal}
