% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{.circe_demographics_to_filters}
\alias{.circe_demographics_to_filters}
\title{Convert Circe demographic criteria back into recipe filters}
\usage{
.circe_demographics_to_filters(demo_list)
}
\description{
Convert Circe demographic criteria back into recipe filters
}
\keyword{internal}
