% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{.circe_group_from_or}
\alias{.circe_group_from_or}
\title{Convert a recipe OR filter group into a nested Circe CriteriaGroup (Type ANY)}
\usage{
.circe_group_from_or(fg, sets)
}
\description{
Convert a recipe OR filter group into a nested Circe CriteriaGroup (Type ANY)
}
\keyword{internal}
