% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{.circe_ids_for_codeset}
\alias{.circe_ids_for_codeset}
\title{Resolve a Circe CodesetId to its concept-id vector}
\usage{
.circe_ids_for_codeset(codeset_id, concept_sets)
}
\description{
Resolve a Circe CodesetId to its concept-id vector
}
\keyword{internal}
