% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{.circe_measurement_from_filter}
\alias{.circe_measurement_from_filter}
\title{Build a Circe Measurement criteria with an optional value range}
\usage{
.circe_measurement_from_filter(f, sets)
}
\description{
Build a Circe Measurement criteria with an optional value range
}
\keyword{internal}
