% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{.circe_primary_from_index}
\alias{.circe_primary_from_index}
\title{Convert an explicit recipe index event to Circe PrimaryCriteria}
\usage{
.circe_primary_from_index(index_event, sets)
}
\description{
Convert an explicit recipe index event to Circe PrimaryCriteria
}
\keyword{internal}
