% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{.circe_table_from_domain}
\alias{.circe_table_from_domain}
\title{Reverse the Circe domain map (Circe key -> recipe table)}
\usage{
.circe_table_from_domain(domain)
}
\description{
Reverse the Circe domain map (Circe key -> recipe table)
}
\keyword{internal}
