% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{.circe_window_to_recipe}
\alias{.circe_window_to_recipe}
\title{Translate a Circe StartWindow back to a recipe window}
\usage{
.circe_window_to_recipe(sw)
}
\description{
Translate a Circe StartWindow back to a recipe window
}
\keyword{internal}
