% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio.R
\name{.clean_ds_error}
\alias{.clean_ds_error}
\title{Clean DataSHIELD error messages for UI display}
\usage{
.clean_ds_error(e)
}
\arguments{
\item{e}{A condition object.}
}
\value{
Character; cleaned error message.
}
\description{
Clean DataSHIELD error messages for UI display
}
\keyword{internal}
