% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{.codegen_filter}
\alias{.codegen_filter}
\title{Generate R code string for a single omop_filter}
\usage{
.codegen_filter(f)
}
\arguments{
\item{f}{An \code{omop_filter} object.}
}
\value{
Character; R code that recreates the filter.
}
\description{
Generate R code string for a single omop_filter
}
\keyword{internal}
