% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{.codegen_filter_group}
\alias{.codegen_filter_group}
\title{Generate R code string for a filter group}
\usage{
.codegen_filter_group(fg)
}
\arguments{
\item{fg}{An \code{omop_filter_group} object.}
}
\value{
Character; R code that recreates the filter group.
}
\description{
Generate R code string for a filter group
}
\keyword{internal}
