% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{.collect_pop_filter_items}
\alias{.collect_pop_filter_items}
\title{Collect population-level filter items while preserving groups}
\usage{
.collect_pop_filter_items(recipe)
}
\arguments{
\item{recipe}{An \code{omop_recipe} object.}
}
\value{
List of matching \code{omop_filter} or \code{omop_filter_group}
  objects.
}
\description{
Collect population-level filter items while preserving groups
}
\keyword{internal}
