% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{.collect_row_filters}
\alias{.collect_row_filters}
\title{Collect all row-level filters from a recipe, flattening groups}
\usage{
.collect_row_filters(recipe)
}
\arguments{
\item{recipe}{An \code{omop_recipe} object.}
}
\value{
List of \code{omop_filter} objects at the \code{"row"} level.
}
\description{
Collect all row-level filters from a recipe, flattening groups
}
\keyword{internal}
