% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_concept_picker.R
\name{.concept_multi_picker_server}
\alias{.concept_multi_picker_server}
\title{Studio Multi-Concept Picker Server}
\usage{
.concept_multi_picker_server(id, state)
}
\arguments{
\item{id}{Character; Shiny module namespace ID.}

\item{state}{Reactive values; the shared OMOP session state.}
}
\value{
A reactive returning integer vector of selected concept IDs.
}
\description{
Studio Multi-Concept Picker Server
}
\keyword{internal}
