% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_concept_picker.R
\name{.concept_multi_picker_ui}
\alias{.concept_multi_picker_ui}
\title{Studio Multi-Concept Picker UI}
\usage{
.concept_multi_picker_ui(
  id,
  label = "Concepts",
  placeholder = "Search to add concepts..."
)
}
\arguments{
\item{id}{Character; Shiny module namespace ID.}

\item{label}{Character; input label.}

\item{placeholder}{Character; placeholder text.}
}
\value{
A Shiny UI element.
}
\description{
Studio Multi-Concept Picker UI
}
\keyword{internal}
