% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_concept_picker.R
\name{.concept_picker_ui}
\alias{.concept_picker_ui}
\title{Studio Concept Picker UI}
\usage{
.concept_picker_ui(
  id,
  label = "Search Concept",
  placeholder = "Type name or ID...",
  multiple = FALSE
)
}
\arguments{
\item{id}{Character; Shiny module namespace ID.}

\item{label}{Character; input label.}

\item{placeholder}{Character; placeholder text.}

\item{multiple}{Logical; allow multiple selections.}
}
\value{
A Shiny UI element.
}
\description{
Studio Concept Picker UI
}
\keyword{internal}
