% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_explore.R
\name{.cs_pick_scope}
\alias{.cs_pick_scope}
\title{Pick the pooled element of a per_site/pooled result, else the first site}
\usage{
.cs_pick_scope(result, scope = "pooled")
}
\description{
Pick the pooled element of a per_site/pooled result, else the first site
}
\keyword{internal}
