% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.dsomop_client_env}
\alias{.dsomop_client_env}
\title{Internal environment for storing dsOMOPClient session state}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.dsomop_client_env
}
\description{
Internal environment for storing dsOMOPClient session state
}
\keyword{internal}
