% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{.flatten_filters}
\alias{.flatten_filters}
\title{Recursively flatten filter groups into individual filters at a given level}
\usage{
.flatten_filters(filters, level = NULL)
}
\arguments{
\item{filters}{List of \code{omop_filter} and/or \code{omop_filter_group}
objects.}

\item{level}{Character or \code{NULL}; filter level to keep (\code{NULL}
keeps all).}
}
\value{
Flat list of \code{omop_filter} objects.
}
\description{
Recursively flatten filter groups into individual filters at a given level
}
\keyword{internal}
