% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.format_r_value}
\alias{.format_r_value}
\title{Format an R value for code generation}
\usage{
.format_r_value(x)
}
\arguments{
\item{x}{An R value to format as code}
}
\value{
Character string of valid R code
}
\description{
Format an R value for code generation
}
\keyword{internal}
