% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_codegen.R
\name{.highlightR}
\alias{.highlightR}
\title{Simple R syntax highlighting via regex -> HTML spans}
\usage{
.highlightR(code)
}
\arguments{
\item{code}{Character; R source code}
}
\value{
Character; HTML-escaped code with span classes
}
\description{
Simple R syntax highlighting via regex -> HTML spans
}
\keyword{internal}
