% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_atlas.R
\name{.mod_atlas_server}
\alias{.mod_atlas_server}
\title{Studio Achilles Atlas Server}
\usage{
.mod_atlas_server(id, state, parent_session = NULL)
}
\arguments{
\item{id}{Character; Shiny module namespace ID.}

\item{state}{Reactive values; the shared OMOP session state.}

\item{parent_session}{Shiny session; the parent session for tab navigation.}
}
\value{
NULL (Shiny module server, called for side effects).
}
\description{
Studio Achilles Atlas Server
}
\keyword{internal}
