% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_atlas.R
\name{.mod_atlas_ui}
\alias{.mod_atlas_ui}
\title{Studio Achilles Atlas UI}
\usage{
.mod_atlas_ui(id)
}
\arguments{
\item{id}{Character; Shiny module namespace ID.}
}
\value{
A Shiny UI element.
}
\description{
Studio Achilles Atlas UI
}
\keyword{internal}
