% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_build.R
\name{.mod_build_server}
\alias{.mod_build_server}
\title{Studio Recipe Builder Server}
\usage{
.mod_build_server(id, state)
}
\arguments{
\item{id}{Character; Shiny module namespace ID.}

\item{state}{Reactive values; the shared OMOP session state.}
}
\value{
NULL (Shiny module server, called for side effects).
}
\description{
Studio Recipe Builder Server
}
\keyword{internal}
