% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_build.R
\name{.mod_build_ui}
\alias{.mod_build_ui}
\title{Studio Recipe Builder UI}
\usage{
.mod_build_ui(id)
}
\arguments{
\item{id}{Character; Shiny module namespace ID.}
}
\value{
A Shiny UI element.
}
\description{
Studio Recipe Builder UI
}
\keyword{internal}
