% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_explore.R
\name{.mod_explore_ui}
\alias{.mod_explore_ui}
\title{Studio Data Explorer UI}
\usage{
.mod_explore_ui(id)
}
\arguments{
\item{id}{Character; Shiny module namespace ID.}
}
\value{
A Shiny UI element.
}
\description{
Studio Data Explorer UI
}
\keyword{internal}
