% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_ohdsi.R
\name{.mod_ohdsi_results_ui}
\alias{.mod_ohdsi_results_ui}
\title{Studio OHDSI Results UI}
\usage{
.mod_ohdsi_results_ui(id)
}
\arguments{
\item{id}{Character; Shiny module namespace ID.}
}
\value{
A Shiny UI element.
}
\description{
Studio OHDSI Results UI
}
\keyword{internal}
