% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_overview.R
\name{.mod_overview_server}
\alias{.mod_overview_server}
\title{Studio Overview Server}
\usage{
.mod_overview_server(id, state)
}
\arguments{
\item{id}{Character; Shiny module namespace ID.}

\item{state}{Reactive values; the shared OMOP session state.}
}
\value{
NULL (Shiny module server, called for side effects).
}
\description{
Studio Overview Server
}
\keyword{internal}
