% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_overview.R
\name{.mod_overview_ui}
\alias{.mod_overview_ui}
\title{Studio Overview UI}
\usage{
.mod_overview_ui(id)
}
\arguments{
\item{id}{Character; Shiny module namespace ID.}
}
\value{
A Shiny UI element.
}
\description{
Studio Overview UI
}
\keyword{internal}
