% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_queries.R
\name{.mod_queries_server}
\alias{.mod_queries_server}
\title{Studio Query Library Server}
\usage{
.mod_queries_server(id, state)
}
\arguments{
\item{id}{Character; Shiny module namespace ID.}

\item{state}{Reactive values; the shared OMOP session state.}
}
\value{
NULL (Shiny module server, called for side effects).
}
\description{
Studio Query Library Server
}
\keyword{internal}
