% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_queries.R
\name{.mod_queries_ui}
\alias{.mod_queries_ui}
\title{Studio Query Library UI}
\usage{
.mod_queries_ui(id)
}
\arguments{
\item{id}{Character; Shiny module namespace ID.}
}
\value{
A Shiny UI element.
}
\description{
Studio Query Library UI
}
\keyword{internal}
