% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_execute.R
\name{.mod_session_server}
\alias{.mod_session_server}
\title{Studio Session Log Server}
\usage{
.mod_session_server(id, state)
}
\arguments{
\item{id}{Character; Shiny module namespace ID.}

\item{state}{Reactive values; the shared OMOP session state.}
}
\value{
NULL (Shiny module server, called for side effects).
}
\description{
Studio Session Log Server
}
\keyword{internal}
