% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_execute.R
\name{.mod_session_ui}
\alias{.mod_session_ui}
\title{Studio Session Log UI}
\usage{
.mod_session_ui(id)
}
\arguments{
\item{id}{Character; Shiny module namespace ID.}
}
\value{
A Shiny UI element.
}
\description{
Studio Session Log UI
}
\keyword{internal}
