% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_settings.R
\name{.mod_settings_server}
\alias{.mod_settings_server}
\title{Studio Settings Server}
\usage{
.mod_settings_server(id, state)
}
\arguments{
\item{id}{Character; Shiny module namespace ID.}

\item{state}{Reactive values; the shared OMOP session state.}
}
\value{
NULL (Shiny module server, called for side effects).
}
\description{
Studio Settings Server
}
\keyword{internal}
