% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_settings.R
\name{.mod_settings_ui}
\alias{.mod_settings_ui}
\title{Studio Settings UI}
\usage{
.mod_settings_ui(id)
}
\arguments{
\item{id}{Character; Shiny module namespace ID.}
}
\value{
A Shiny UI element.
}
\description{
Studio Settings UI
}
\keyword{internal}
