% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_sidebar.R
\name{.mod_sidebar_ui}
\alias{.mod_sidebar_ui}
\title{Studio Sidebar UI}
\usage{
.mod_sidebar_ui(id)
}
\arguments{
\item{id}{Character; Shiny module namespace ID.}
}
\value{
A Shiny sidebar element.
}
\description{
Studio Sidebar UI
}
\keyword{internal}
