% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_ohdsi.R
\name{.ohdsi_forest_plot}
\alias{.ohdsi_forest_plot}
\title{Forest plot helper for CM, SCCS, Evidence Synthesis}
\usage{
.ohdsi_forest_plot(
  df,
  estimate_col = "rr",
  lower_col = "ci_95_lb",
  upper_col = "ci_95_ub",
  title = "Forest Plot"
)
}
\description{
Forest plot helper for CM, SCCS, Evidence Synthesis
}
\keyword{internal}
