% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_plan_dag.R
\name{.plan_html_escape}
\alias{.plan_html_escape}
\title{HTML-escape a scalar for use inside a visNetwork tooltip}
\usage{
.plan_html_escape(x)
}
\description{
HTML-escape a scalar for use inside a visNetwork tooltip
}
\keyword{internal}
