% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\name{.pool_crosstab_slice}
\alias{.pool_crosstab_slice}
\title{Pool a single (unstratified) set of cross-tab slices}
\usage{
.pool_crosstab_slice(slices, policy = "strict")
}
\description{
Pool a single (unstratified) set of cross-tab slices
}
\keyword{internal}
