% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{.schema_time_window}
\alias{.schema_time_window}
\title{Render a variable time_window as a preview string}
\usage{
.schema_time_window(tw)
}
\description{
Render a variable time_window as a preview string
}
\keyword{internal}
