% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_codegen.R
\name{.studio_codegen_concept_set}
\alias{.studio_codegen_concept_set}
\title{Generate R code for a concept set}
\usage{
.studio_codegen_concept_set(concept_ids, name = "my_concepts")
}
\arguments{
\item{concept_ids}{Integer vector}

\item{name}{Character; variable name}
}
\value{
Character; R code
}
\description{
Generate R code for a concept set
}
\keyword{internal}
