% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_codegen.R
\name{.studio_codegen_exploration}
\alias{.studio_codegen_exploration}
\title{Generate R code for an exploration function call}
\usage{
.studio_codegen_exploration(fn_name, args_list)
}
\arguments{
\item{fn_name}{Character; function name (e.g. "ds.omop.concept.prevalence")}

\item{args_list}{Named list of arguments}
}
\value{
Character; R code string
}
\description{
Generate R code for an exploration function call
}
\keyword{internal}
