% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_codegen.R
\name{.studio_codegen_plan}
\alias{.studio_codegen_plan}
\title{Generate R code for a plan}
\usage{
.studio_codegen_plan(plan, out, symbol = "omop")
}
\arguments{
\item{plan}{An omop_plan object}

\item{out}{Named character vector; output symbol mapping}

\item{symbol}{Character; OMOP session symbol}
}
\value{
Character; reproducible R script
}
\description{
Generate R code for a plan
}
\keyword{internal}
