% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_codegen.R
\name{.studio_codegen_vocabulary}
\alias{.studio_codegen_vocabulary}
\title{Generate R code for a vocabulary function call}
\usage{
.studio_codegen_vocabulary(fn_name, args_list)
}
\arguments{
\item{fn_name}{Character; function name}

\item{args_list}{Named list of arguments}
}
\value{
Character; R code string
}
\description{
Generate R code for a vocabulary function call
}
\keyword{internal}
