% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio.R
\name{.studio_server}
\alias{.studio_server}
\title{OMOP Studio Server}
\usage{
.studio_server(symbol, captured_session = NULL)
}
\arguments{
\item{symbol}{Character; OMOP session symbol.}

\item{captured_session}{Captured session object (protection against namespace reloads).}
}
\value{
A Shiny server function.
}
\description{
OMOP Studio Server
}
\keyword{internal}
