% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio.R
\name{.studio_ui}
\alias{.studio_ui}
\title{OMOP Studio UI}
\usage{
.studio_ui(symbol)
}
\arguments{
\item{symbol}{Character; OMOP session symbol.}
}
\value{
A function returning a Shiny UI definition.
}
\description{
OMOP Studio UI
}
\keyword{internal}
