% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{.validate_plan_harmonization}
\alias{.validate_plan_harmonization}
\title{Validate a plan against an already negotiated federation contract}
\usage{
.validate_plan_harmonization(plan, contract)
}
\description{
Validate a plan against an already negotiated federation contract
}
\keyword{internal}
