% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{ds.omop.code}
\alias{ds.omop.code}
\title{Extract reproducible R code from a dsomop_result}
\usage{
ds.omop.code(x)
}
\arguments{
\item{x}{A dsomop_result object}
}
\value{
Character string of R code
}
\description{
Extract reproducible R code from a dsomop_result
}
