% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{ds.omop.columns}
\alias{ds.omop.columns}
\title{List columns in a table}
\usage{
ds.omop.columns(table, symbol = "omop", conns = NULL)
}
\arguments{
\item{table}{Character; table name}

\item{symbol}{Character; OMOP session symbol}

\item{conns}{DSI connections}
}
\value{
Named list (per server) of data frames with column metadata
}
\description{
List columns in a table
}
