% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary.R
\name{ds.omop.concept.expand}
\alias{ds.omop.concept.expand}
\title{Expand a concept set to full ID list}
\usage{
ds.omop.concept.expand(
  concept_set,
  symbol = "omop",
  conns = NULL,
  execute = TRUE
)
}
\arguments{
\item{concept_set}{An omop_concept_set object}

\item{symbol}{Character; OMOP session symbol}

\item{conns}{DSI connections}

\item{execute}{Logical; if FALSE, return dry-run result with code only}
}
\value{
A dsomop_result object
}
\description{
Expand a concept set to full ID list
}
