% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{ds.omop.copy_code}
\alias{ds.omop.copy_code}
\title{Copy reproducible R code to clipboard}
\usage{
ds.omop.copy_code(x)
}
\arguments{
\item{x}{A dsomop_result object}
}
\value{
Invisibly returns the code string
}
\description{
Copy reproducible R code to clipboard
}
