% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-resource.R
\name{generateResourceSymbol}
\alias{generateResourceSymbol}
\title{Generate a Unique Resource Symbol for an OMOP CDM Database Resource}
\usage{
generateResourceSymbol(resource)
}
\arguments{
\item{resource}{A character string representing the resource for which the symbol is generated.}
}
\value{
A character string representing the unique symbol for the resource.
}
\description{
This function generates a unique symbol name for a connection resource to an OMOP CDM database.
The generated symbol will be stored in the OMOPCDMDatabase object. It includes a random string
to prevent overwriting other objects in the DataSHIELD environment where the user is working.
}
