% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{omop.feature.boolean}
\alias{omop.feature.boolean}
\title{Binary presence/absence feature}
\usage{
omop.feature.boolean(concept_set, name = NULL)
}
\arguments{
\item{concept_set}{Numeric vector of concept IDs or omop_concept_set}

\item{name}{Character; feature name}
}
\value{
An omop_feature_spec list
}
\description{
Binary presence/absence feature
}
