% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{omop.feature.count}
\alias{omop.feature.count}
\title{Count of events per person}
\usage{
omop.feature.count(concept_set, name = NULL)
}
\arguments{
\item{concept_set}{Numeric vector of concept IDs or omop_concept_set}

\item{name}{Character; feature name}
}
\value{
An omop_feature_spec list
}
\description{
Count of events per person
}
