% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{omop.feature.first_value}
\alias{omop.feature.first_value}
\title{First recorded value feature}
\usage{
omop.feature.first_value(
  concept_set,
  value_column = "value_as_number",
  name = NULL
)
}
\arguments{
\item{concept_set}{Numeric vector or omop_concept_set}

\item{value_column}{Character; column to extract value from}

\item{name}{Character; feature name}
}
\value{
An omop_feature_spec list
}
\description{
First recorded value feature
}
