% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-concepts.R
\name{OMOPCDMDatabase-concepts}
\alias{OMOPCDMDatabase-concepts}
\title{Get Concepts Catalog}
\arguments{
\item{table}{The name of the table for which to retrieve the concepts catalog.}
}
\value{
A data frame containing concept_id and concept_name columns representing available concepts in the specified
table of the OMOP CDM database.
}
\description{
This function returns a data frame of available concept IDs and their names for a specified table in an OMOP CDM database.
}
