% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_resource.R
\name{OMOPCDMDatabase-removeResource}
\alias{OMOPCDMDatabase-removeResource}
\title{Remove the OMOP CDM database resource from a DataSHIELD environment}
\description{
This function removes the resource of the OMOP CDM database connection
from the DataSHIELD environment. It is commonly used after having performed operations
related to the connection with the database.
}
