% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy.R
\name{print.omop_dp_status}
\alias{print.omop_dp_status}
\title{Print the release policy on each DP server}
\usage{
\method{print}{omop_dp_status}(x, ...)
}
\arguments{
\item{x}{An \code{omop_dp_status} returned by
\code{\link{ds.omop.dp.status}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print the release policy on each DP server
}
