% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{print.omop_filter_group}
\alias{print.omop_filter_group}
\title{Print an omop_filter_group}
\usage{
\method{print}{omop_filter_group}(x, ...)
}
\arguments{
\item{x}{An \code{omop_filter_group} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Print an omop_filter_group
}
