% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{print.omop_output}
\alias{print.omop_output}
\title{Print an omop_output}
\usage{
\method{print}{omop_output}(x, ...)
}
\arguments{
\item{x}{An \code{omop_output} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Print an omop_output
}
