% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{print.omop_plan}
\alias{print.omop_plan}
\title{Print method for extraction plans}
\usage{
\method{print}{omop_plan}(x, ...)
}
\arguments{
\item{x}{An omop_plan object}

\item{...}{Additional arguments}
}
\description{
Print method for extraction plans
}
