% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{print.omop_population}
\alias{print.omop_population}
\title{Print an omop_population}
\usage{
\method{print}{omop_population}(x, ...)
}
\arguments{
\item{x}{An \code{omop_population} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Print an omop_population
}
