% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{print.omop_recipe}
\alias{print.omop_recipe}
\title{Print an omop_recipe}
\usage{
\method{print}{omop_recipe}(x, ...)
}
\arguments{
\item{x}{An \code{omop_recipe} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Print an omop_recipe
}
