% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{print.omop_variable}
\alias{print.omop_variable}
\title{Print an omop_variable}
\usage{
\method{print}{omop_variable}(x, ...)
}
\arguments{
\item{x}{An \code{omop_variable} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Print an omop_variable
}
