% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{print.omop_variable_block}
\alias{print.omop_variable_block}
\title{Print an omop_variable_block}
\usage{
\method{print}{omop_variable_block}(x, ...)
}
\arguments{
\item{x}{An \code{omop_variable_block} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Print an omop_variable_block
}
