% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{recipe_add_output}
\alias{recipe_add_output}
\title{Add an output specification to the recipe}
\usage{
recipe_add_output(recipe, output)
}
\arguments{
\item{recipe}{An \code{omop_recipe} object.}

\item{output}{An \code{\link{omop_output}} object.}
}
\value{
The modified \code{omop_recipe} object.
}
\description{
Registers an output specification that defines how extracted data should be
shaped. Multiple outputs can target different subsets of variables and
populations, each producing a separate server-side dataset.
}
\seealso{
\code{\link{omop_output}}, \code{\link{omop_recipe}}
}
\keyword{internal}
