% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{recipe_add_variable}
\alias{recipe_add_variable}
\title{Add a variable to the recipe}
\usage{
recipe_add_variable(recipe, variable = NULL, ...)
}
\arguments{
\item{recipe}{An omop_recipe}

\item{variable}{An omop_variable object, or NULL to construct from ...}

\item{...}{If variable is NULL, arguments passed to omop_variable()}
}
\value{
The modified recipe
}
\description{
Add a variable to the recipe
}
