% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{recipe_clear}
\alias{recipe_clear}
\title{Clear the entire recipe}
\usage{
recipe_clear(recipe)
}
\arguments{
\item{recipe}{An omop_recipe}
}
\value{
An empty recipe
}
\description{
Clear the entire recipe
}
