% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{recipe_remove_filter}
\alias{recipe_remove_filter}
\title{Remove a filter from the recipe}
\usage{
recipe_remove_filter(recipe, id)
}
\arguments{
\item{recipe}{An omop_recipe}

\item{id}{Character; filter ID to remove}
}
\value{
The modified recipe
}
\description{
Remove a filter from the recipe
}
