% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{recipe_remove_output}
\alias{recipe_remove_output}
\title{Remove an output specification from the recipe}
\usage{
recipe_remove_output(recipe, name)
}
\arguments{
\item{recipe}{An omop_recipe}

\item{name}{Character; output name to remove}
}
\value{
The modified recipe
}
\description{
Remove an output specification from the recipe
}
