% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{recipe_remove_population}
\alias{recipe_remove_population}
\title{Remove a population from the recipe}
\usage{
recipe_remove_population(recipe, id)
}
\arguments{
\item{recipe}{An omop_recipe}

\item{id}{Character; population ID to remove}
}
\value{
The modified recipe
}
\description{
Remove a population from the recipe
}
