% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{recipe_remove_variable}
\alias{recipe_remove_variable}
\title{Remove a variable from the recipe}
\usage{
recipe_remove_variable(recipe, name)
}
\arguments{
\item{recipe}{An \code{omop_recipe} object.}

\item{name}{Character; variable name to remove.}
}
\value{
The modified \code{omop_recipe} object.
}
\description{
Removes a variable by name from the recipe's variable list.
}
\examples{
\dontrun{
recipe <- recipe_remove_variable(recipe, "condition_occurrence_c201820")
}
}
\seealso{
\code{\link{recipe_add_variable}}
}
