% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{as.data.frame.dsomop_result}
\alias{as.data.frame.dsomop_result}
\title{Convert dsomop_result to data.frame}
\usage{
\method{as.data.frame}{dsomop_result}(x, ...)
}
\arguments{
\item{x}{A dsomop_result object}

\item{...}{Ignored}
}
\value{
A data.frame
}
\description{
Returns pooled result if available, otherwise the first server's result.
}
