% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{.compile_filter_node}
\alias{.compile_filter_node}
\title{Compile a single filter or filter group into the server filter DSL}
\usage{
.compile_filter_node(f)
}
\description{
Compile a single filter or filter group into the server filter DSL
}
\keyword{internal}
