% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{.compile_filter_tree}
\alias{.compile_filter_tree}
\title{Compile a list of filters into an AND/OR tree for the server filter DSL}
\usage{
.compile_filter_tree(filters, default_operator = "and")
}
\description{
Compile a list of filters into an AND/OR tree for the server filter DSL
}
\keyword{internal}
