% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{.extract_filters_by_level}
\alias{.extract_filters_by_level}
\title{Extract filters (or groups containing filters) at a given level}
\usage{
.extract_filters_by_level(filters, level)
}
\description{
Extract filters (or groups containing filters) at a given level
}
\keyword{internal}
