% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio.R
\name{.get_person_tables}
\alias{.get_person_tables}
\title{Extract CDM table names with person_id from state tables}
\usage{
.get_person_tables(tables)
}
\arguments{
\item{tables}{Named list of table metadata per server.}
}
\value{
Character vector of table names.
}
\description{
Extract CDM table names with person_id from state tables
}
\keyword{internal}
