% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{.group_has_level}
\alias{.group_has_level}
\title{Check if a filter group contains any filter at a given level}
\usage{
.group_has_level(group, level)
}
\description{
Check if a filter group contains any filter at a given level
}
\keyword{internal}
