% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{ds.omop.cohort.combine}
\alias{ds.omop.cohort.combine}
\title{Combine two cohorts with set operations}
\usage{
ds.omop.cohort.combine(
  op,
  cohort_a,
  cohort_b,
  new_name = NULL,
  symbol = "omop",
  conns = NULL
)
}
\arguments{
\item{op}{Character; "intersect", "union", or "setdiff"}

\item{cohort_a}{Character; first cohort temp table}

\item{cohort_b}{Character; second cohort temp table}

\item{new_name}{Character; name for result}

\item{symbol}{Character; OMOP session symbol}

\item{conns}{DSI connections}
}
\value{
Invisible; combined cohort reference
}
\description{
Combine two cohorts with set operations
}
