% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{ds.omop.cohort.definition}
\alias{ds.omop.cohort.definition}
\title{Get a cohort definition by ID}
\usage{
ds.omop.cohort.definition(id, symbol = "omop", conns = NULL)
}
\arguments{
\item{id}{Integer; cohort definition ID}

\item{symbol}{Character; OMOP session symbol}

\item{conns}{DSI connections}
}
\value{
Named list (per server) with definition details
}
\description{
Get a cohort definition by ID
}
