% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{ds.omop.cohort.list}
\alias{ds.omop.cohort.list}
\title{List available cohort definitions}
\usage{
ds.omop.cohort.list(symbol = "omop", conns = NULL)
}
\arguments{
\item{symbol}{Character; OMOP session symbol}

\item{conns}{DSI connections}
}
\value{
Named list (per server) of data frames
}
\description{
List available cohort definitions
}
