% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{ds.omop.cohort.ref}
\alias{ds.omop.cohort.ref}
\title{Create a cohort reference for the plan DSL (client-only)}
\usage{
ds.omop.cohort.ref(cohort_definition_id)
}
\arguments{
\item{cohort_definition_id}{Integer; ID of existing cohort}
}
\value{
List; a cohort reference for plans
}
\description{
Create a cohort reference for the plan DSL (client-only)
}
