% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{ds.omop.compare}
\alias{ds.omop.compare}
\title{Compare schemas across servers}
\usage{
ds.omop.compare(symbol = "omop", conns = NULL)
}
\arguments{
\item{symbol}{Character; OMOP session symbol}

\item{conns}{DSI connections}
}
\value{
List with common_tables, server_only, column_diffs
}
\description{
Compare schemas across servers
}
