% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{ds.omop.disconnect}
\alias{ds.omop.disconnect}
\title{Disconnect an OMOP session}
\usage{
ds.omop.disconnect(symbol = "omop", conns = NULL)
}
\arguments{
\item{symbol}{Character; the session symbol to disconnect}

\item{conns}{DSI connections}
}
\description{
Disconnect an OMOP session
}
