% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{ds.omop.joins}
\alias{ds.omop.joins}
\title{Get the join relationship graph}
\usage{
ds.omop.joins(symbol = "omop", conns = NULL)
}
\arguments{
\item{symbol}{Character; OMOP session symbol}

\item{conns}{DSI connections}
}
\value{
Named list (per server) of data frames
}
\description{
Get the join relationship graph
}
