% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{ds.omop.plan.cohort}
\alias{ds.omop.plan.cohort}
\title{Set a cohort filter on the plan}
\usage{
ds.omop.plan.cohort(plan, cohort_definition_id = NULL, spec = NULL)
}
\arguments{
\item{plan}{An omop_plan object}

\item{cohort_definition_id}{Integer; ID of existing cohort}

\item{spec}{Named list; cohort specification DSL}
}
\value{
The modified plan
}
\description{
Set a cohort filter on the plan
}
