% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{ds.omop.plan.features}
\alias{ds.omop.plan.features}
\title{Add feature extraction with recipe specs}
\usage{
ds.omop.plan.features(plan, name, table, specs)
}
\arguments{
\item{plan}{An omop_plan object}

\item{name}{Character; output name}

\item{table}{Character; source table name}

\item{specs}{Named list of omop_feature_spec objects}
}
\value{
The modified plan
}
\description{
Add feature extraction with recipe specs
}
