% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{ds.omop.plan.harmonize}
\alias{ds.omop.plan.harmonize}
\title{Harmonize a plan for multi-server execution}
\usage{
ds.omop.plan.harmonize(
  plan,
  mode = "intersection",
  symbol = "omop",
  conns = NULL
)
}
\arguments{
\item{plan}{An omop_plan object}

\item{mode}{Character; "intersection" or "union_with_missing"}

\item{symbol}{Character; OMOP session symbol}

\item{conns}{DSI connections}
}
\value{
The harmonized plan
}
\description{
Harmonize a plan for multi-server execution
}
