% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{ds.omop.plan.options}
\alias{ds.omop.plan.options}
\title{Set plan-wide options}
\usage{
ds.omop.plan.options(
  plan,
  translate_concepts = NULL,
  block_sensitive = NULL,
  min_persons = NULL
)
}
\arguments{
\item{plan}{An omop_plan object}

\item{translate_concepts}{Logical}

\item{block_sensitive}{Logical}

\item{min_persons}{Integer}
}
\value{
The modified plan
}
\description{
Set plan-wide options
}
