% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{ds.omop.plan.outcome}
\alias{ds.omop.plan.outcome}
\title{Add an outcome extraction (convenience)}
\usage{
ds.omop.plan.outcome(plan, name, concept_set, table = "condition_occurrence")
}
\arguments{
\item{plan}{An omop_plan object}

\item{name}{Character; output name}

\item{concept_set}{Numeric vector; outcome concept IDs}

\item{table}{Character; source table}
}
\value{
The modified plan
}
\description{
Add an outcome extraction (convenience)
}
