% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{ds.omop.plan.preview}
\alias{ds.omop.plan.preview}
\title{Preview a plan (safe aggregate)}
\usage{
ds.omop.plan.preview(plan, symbol = "omop", conns = NULL)
}
\arguments{
\item{plan}{An omop_plan object}

\item{symbol}{Character; OMOP session symbol}

\item{conns}{DSI connections}
}
\value{
Named list (per server) with preview info
}
\description{
Preview a plan (safe aggregate)
}
