% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{ds.omop.plan.validate}
\alias{ds.omop.plan.validate}
\title{Validate a plan against server schemas}
\usage{
ds.omop.plan.validate(plan, symbol = "omop", conns = NULL)
}
\arguments{
\item{plan}{An omop_plan object}

\item{symbol}{Character; OMOP session symbol}

\item{conns}{DSI connections}
}
\value{
Named list (per server) with validation results
}
\description{
Validate a plan against server schemas
}
