% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class.R
\name{ds.omop}
\alias{ds.omop}
\title{Factory function for OMOP CDM Database}
\usage{
ds.omop(connections, resource)
}
\arguments{
\item{connections}{Connection object to the DataSHIELD server.}

\item{resource}{Identifier for the specific resource within the DataSHIELD server.}
}
\value{
A new instance of the OMOPCDMDatabase class.
}
\description{
This function creates a new instance of the OMOPCDMDatabase class, allowing
for interaction with an OMOP CDM database.
}
