% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{ds.omop.snapshot}
\alias{ds.omop.snapshot}
\title{Get a full schema snapshot}
\usage{
ds.omop.snapshot(symbol = "omop", conns = NULL)
}
\arguments{
\item{symbol}{Character; OMOP session symbol}

\item{conns}{DSI connections}
}
\value{
Named list (per server) with tables, cdm_info, edges
}
\description{
Get a full schema snapshot
}
