% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{ds.omop.status}
\alias{ds.omop.status}
\title{Get OMOP session status}
\usage{
ds.omop.status(symbol = "omop")
}
\arguments{
\item{symbol}{Character; session symbol}
}
\value{
Named list with per-server status
}
\description{
Get OMOP session status
}
