% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{ds.omop.tables}
\alias{ds.omop.tables}
\title{List tables in the OMOP CDM database}
\usage{
ds.omop.tables(schema_category = NULL, symbol = "omop", conns = NULL)
}
\arguments{
\item{schema_category}{Character; filter by "CDM", "Vocabulary", "Results"}

\item{symbol}{Character; OMOP session symbol}

\item{conns}{DSI connections}
}
\value{
Named list (per server) of data frames with table metadata
}
\description{
List tables in the OMOP CDM database
}
