% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_get.R
\name{getTableCall}
\alias{getTableCall}
\title{Construct the Call for Table Assignment}
\usage{
getTableCall(
  resource,
  table,
  conceptFilter = NULL,
  columnFilter = NULL,
  personFilter = NULL,
  mergeColumn = NULL,
  dropNA = FALSE
)
}
\arguments{
\item{resource}{The resource symbol representing the connection to the OMOP CDM database.}

\item{table}{A string specifying the name of the table to be assigned.}

\item{conceptFilter}{A numeric vector specifying concept IDs to filter the table by.}

\item{columnFilter}{A string vector specifying column names to filter (select) in the table.}

\item{personFilter}{A string specifying the symbol in the environment of a table from which to obtain person IDs.}

\item{mergeColumn}{A string specifying the column name for merging operations with other tables.
Defaults to "person_id" if not specified.}

\item{dropNA}{A boolean indicating whether to drop empty columns. Defaults to FALSE.}
}
\value{
A string representing the call to be executed on the DataSHIELD server for table assignment.
}
\description{
This function assists in constructing the call that will be sent from the DataSHIELD client to the server
to perform the assignment of a table from an OMOP CDM database based on the parameters specified by the user.
}
