% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{omop.feature.time_since}
\alias{omop.feature.time_since}
\title{Time since event feature}
\usage{
omop.feature.time_since(
  concept_set,
  reference_date = NULL,
  unit = "day",
  name = NULL
)
}
\arguments{
\item{concept_set}{Numeric vector or omop_concept_set}

\item{reference_date}{Character; reference date (ISO format)}

\item{unit}{Character; "day" or "month"}

\item{name}{Character; feature name}
}
\value{
An omop_feature_spec list
}
\description{
Time since event feature
}
