% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_resource.R
\name{OMOPCDMDatabase-assignResource}
\alias{OMOPCDMDatabase-assignResource}
\title{Assign the OMOP CDM database resource to a DataSHIELD environment}
\arguments{
\item{symbol}{A character string representing the symbol to which the
resource will be assigned.}
}
\description{
This function assigns the resource of the OMOP CDM database connection
to the DataSHIELD environment, facilitating the connection to the database from the
server side.
}
