% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-check.R
\name{OMOPCDMDatabase-checkConnection}
\alias{OMOPCDMDatabase-checkConnection}
\title{Check Database Connection}
\value{
A boolean value indicating whether the connection to the database is valid (TRUE) or not (FALSE).
}
\description{
This function checks if the OMOPCDMDatabase object points to a valid connection to an OMOP CDM database.
}
