% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_check.R
\name{OMOPCDMDatabase-checkPrivacyControlLevel}
\alias{OMOPCDMDatabase-checkPrivacyControlLevel}
\title{Check Privacy Control Level}
\value{
A list of warnings for each server where the privacy control level may not be permissive enough.
Each element of the list is named by the server and contains the warning message.
}
\description{
This function checks if the privacy control level settings of the server configuration
allow the use of dsOMOP operations.
}
