% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-columns.R
\name{OMOPCDMDatabase-columns}
\alias{OMOPCDMDatabase-columns}
\title{Get Columns Catalog}
\arguments{
\item{table}{The name of the table for which to retrieve the columns catalog.}

\item{dropNA}{A boolean indicating whether to skip empty columns.}
}
\value{
A list of column names available in the specified table of the OMOP CDM database.
}
\description{
This function returns a list of available columns for a specified table in an OMOP CDM database.
}
