% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_tables.R
\name{OMOPCDMDatabase-tables}
\alias{OMOPCDMDatabase-tables}
\title{Get Tables Catalog}
\value{
A list of table names available in the OMOP CDM database.
}
\description{
This function returns a list of available tables in an OMOP CDM database.
}
