% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-class.R
\name{OMOPCDMDatabase}
\alias{OMOPCDMDatabase}
\title{OMOP CDM Database class}
\description{
This class represents a connection to an OMOP CDM database, providing
various methods for the user to interact with the database. It allows for the retrieval
of tables as well as fetching catalogs of available data within the database.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OMOPCDMDatabase-new}{\code{OMOPCDMDatabase$new()}}
\item \href{#method-OMOPCDMDatabase-assignResource}{\code{OMOPCDMDatabase$assignResource()}}
\item \href{#method-OMOPCDMDatabase-removeResource}{\code{OMOPCDMDatabase$removeResource()}}
\item \href{#method-OMOPCDMDatabase-get}{\code{OMOPCDMDatabase$get()}}
\item \href{#method-OMOPCDMDatabase-tables}{\code{OMOPCDMDatabase$tables()}}
\item \href{#method-OMOPCDMDatabase-columns}{\code{OMOPCDMDatabase$columns()}}
\item \href{#method-OMOPCDMDatabase-concepts}{\code{OMOPCDMDatabase$concepts()}}
\item \href{#method-OMOPCDMDatabase-checkConnection}{\code{OMOPCDMDatabase$checkConnection()}}
\item \href{#method-OMOPCDMDatabase-clone}{\code{OMOPCDMDatabase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMDatabase-new"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMDatabase-new}{}}}
\subsection{Method \code{new()}}{
This method initializes a new instance of the OMOPCDMDatabase class.
It sets up the connection to the OMOP CDM database given the provided
DataSHIELD connection object and the OMOP CDM database resource identifier.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMDatabase$new(connections, resource)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connections}}{Connection object to the DataSHIELD server.}

\item{\code{resource}}{Identifier for the specific resource within the DataSHIELD server.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the OMOPCDMDatabase class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMDatabase-assignResource"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMDatabase-assignResource}{}}}
\subsection{Method \code{assignResource()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMDatabase$assignResource(symbol)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMDatabase-removeResource"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMDatabase-removeResource}{}}}
\subsection{Method \code{removeResource()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMDatabase$removeResource()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMDatabase-get"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMDatabase-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMDatabase$get(
  table,
  symbol = NULL,
  conceptFilter = NULL,
  columnFilter = NULL,
  personFilter = NULL,
  mergeColumn = NULL,
  dropNA = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMDatabase-tables"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMDatabase-tables}{}}}
\subsection{Method \code{tables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMDatabase$tables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMDatabase-columns"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMDatabase-columns}{}}}
\subsection{Method \code{columns()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMDatabase$columns(table, dropNA = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMDatabase-concepts"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMDatabase-concepts}{}}}
\subsection{Method \code{concepts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMDatabase$concepts(table)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMDatabase-checkConnection"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMDatabase-checkConnection}{}}}
\subsection{Method \code{checkConnection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMDatabase$checkConnection()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMDatabase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMDatabase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMDatabase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
