% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{print.dsomop_result}
\alias{print.dsomop_result}
\title{Print a dsomop_result}
\usage{
\method{print}{dsomop_result}(x, ...)
}
\arguments{
\item{x}{A dsomop_result object}

\item{...}{Ignored}
}
\description{
Print a dsomop_result
}
