% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{recipe_add_block}
\alias{recipe_add_block}
\title{Add a variable block to the recipe}
\usage{
recipe_add_block(recipe, block)
}
\arguments{
\item{recipe}{An omop_recipe}

\item{block}{An omop_variable_block object}
}
\value{
The modified recipe
}
\description{
Expands concept_ids into individual omop_variable objects using the block's
defaults, and ensures unique naming.
}
