% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{recipe_add_filter}
\alias{recipe_add_filter}
\title{Add a filter to the recipe}
\usage{
recipe_add_filter(recipe, filter, id = NULL)
}
\arguments{
\item{recipe}{An omop_recipe}

\item{filter}{An omop_filter or omop_filter_group object}

\item{id}{Character or NULL; filter ID (auto-generated if NULL)}
}
\value{
The modified recipe
}
\description{
Add a filter to the recipe
}
