% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{recipe_add_output}
\alias{recipe_add_output}
\title{Add an output specification to the recipe}
\usage{
recipe_add_output(recipe, output)
}
\arguments{
\item{recipe}{An omop_recipe}

\item{output}{An omop_output object}
}
\value{
The modified recipe
}
\description{
Add an output specification to the recipe
}
