% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{recipe_add_population}
\alias{recipe_add_population}
\title{Add a population to the recipe}
\usage{
recipe_add_population(recipe, population)
}
\arguments{
\item{recipe}{An \code{omop_recipe} object.}

\item{population}{An \code{\link{omop_population}} object to add.}
}
\value{
The modified \code{omop_recipe} object.
}
\description{
Registers a new population node in the recipe's population DAG. The
population's parent (if any) must already exist in the recipe. The base
population is created automatically by \code{\link{omop_recipe}}.
}
\examples{
\dontrun{
recipe <- omop_recipe()
recipe <- recipe_add_population(recipe,
  omop_population(id = "adults", label = "Adults 18+",
                  parent_id = "base",
                  filters = list(omop_filter_age(min = 18))))
}
}
\seealso{
\code{\link{recipe_remove_population}}, \code{\link{omop_population}}
}
