% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{recipe_add_population}
\alias{recipe_add_population}
\title{Add a population to the recipe}
\usage{
recipe_add_population(recipe, population)
}
\arguments{
\item{recipe}{An omop_recipe}

\item{population}{An omop_population object}
}
\value{
The modified recipe
}
\description{
Add a population to the recipe
}
