% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{recipe_import_json}
\alias{recipe_import_json}
\title{Import a recipe from JSON}
\usage{
recipe_import_json(json)
}
\arguments{
\item{json}{Character; JSON string or file path}
}
\value{
An omop_recipe object
}
\description{
Import a recipe from JSON
}
