% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{recipe_remove_output}
\alias{recipe_remove_output}
\title{Remove an output specification from the recipe}
\usage{
recipe_remove_output(recipe, name)
}
\arguments{
\item{recipe}{An \code{omop_recipe} object.}

\item{name}{Character; output name to remove.}
}
\value{
The modified \code{omop_recipe} object.
}
\description{
Removes an output specification by name from the recipe.
}
\examples{
\dontrun{
recipe <- recipe_remove_output(recipe, "wide_data")
}
}
\seealso{
\code{\link{recipe_add_output}}
}
