% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{recipe_remove_variable}
\alias{recipe_remove_variable}
\title{Remove a variable from the recipe}
\usage{
recipe_remove_variable(recipe, name)
}
\arguments{
\item{recipe}{An omop_recipe}

\item{name}{Character; variable name to remove}
}
\value{
The modified recipe
}
\description{
Remove a variable from the recipe
}
