% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{recipe_to_code}
\alias{recipe_to_code}
\title{Generate reproducible R code from a recipe}
\usage{
recipe_to_code(recipe)
}
\arguments{
\item{recipe}{An omop_recipe object}
}
\value{
Character string of R code
}
\description{
Produces minimal R code that recreates the recipe. Does not include
library() calls or header comments.
}
